Add NPM scripts for checking dependencies
This commit adds two new npm scripts:

1. check_dependencies: Will run npm-check-updates and find any
   outdated dependencies. Will exit with status code 1 if anything is
   out of date (-e 2 flag).

2. update_dependencies: Simply runs npm-check-updates with -u flag to
   automatically update anything out of date.

The prepublish script now also calls check_dependencies. That is,
assuming that some CI is running prepublish before publishing, it
will not be possible to publish the package with any out-of-date
deps.

If you want to be less stringent, and allow prepubish to exit
with success even with out-of-date dependencies, you can remove
the -e2 flag. If anything needs updating, it will still report, but
it will exit with 0.