Update digest to explictly use "binary" format when writing to hashes
In node < 5 this was the default for `update` and other crypto methods.
This changed to 'utf8' in https://github.com/nodejs/node/commit/b010c8716498dca398e61c388859fea92296feb3.
This commit enforces the old behavior in our crypto methods and produces
the same digests as node < 5.

I'm not confident that this is the best holistic solution, since this is
not an area of expertise; I'm just getting the tests to pass :)

It's worth checking in with friends who know more about crypto to make
sure that this is an acceptable move.