BTIDEAL-164 - Use correct Cosmos Route ID when generating valid iDEAL Payments (#102)
What
===
Use correct Cosmos Route ID when generating valid iDEAL Payments by
getting it from the configuration for the merchant account.

Why
===
The tests generate iDEAL Payments on the default route of the merchant,
and then attempt to create charges on the merchant account ID that is
setup for iDEAL, but the Gateway cannot guarantee that the route will
be the default route.

When the tests were first written it appears that assumption was made
that the default route would be used.

Also
===
Move the `done()` call in teh iDEAL Payment Transaction test into the
callback of assertions.

If we make it to the `done()` call it signals that the test is complete
and none of the assertions failed. The `done()` call was outside the
callback and so would execute before the assertions would run and so
this test was not really asserting anything.