BTCARDSXP-372 Line item discount and tax amount may be zero
What
===
Update the validations for transaction line item discount amount, tax
amount, and unit tax amount to indicate that a zero value is allowed.

Why
===
The discount amount and tax amount can be zero and we accidentally wrote
a validation that disallowed it. The same change has been made to unit
tax amount for consistency, @abuttars is the significant owner of that
field and he OK'd allowing it there too.

Notes
===
* This relies on a change in the gateway that's coming and so CI will
fail until that is merged to master.
* There's no deprecation of the tax amount field because that field has
never been in a release.