BTCARDSXP-530 Test new transaction source values
### What
Add tests for new transaction source values, `merchant` and
`recurring_first`.

### Why
The values are valid transaction source values to pass to the gateway
and are defined in the developer docs alongside the existing values
`recurring` and `moto`.

### Notes
The values were not added to the `Source` enum because there are two
enumerations returned from the gateway named transaction source, and the
enum represents the status history transaction source.